[Truffle] Made $~ and related global variables frame-local.
Prior to this change we treated $~ as thread-local. While that worked in many cases, it did mean that the match data lived longer than it should. Fixing this is a bit tricky because generally the match data should not propagate up the call stack beyond the caller, but we implement much of our core library in Ruby itself. So we must propagate values back to the user's call, and then cease propagating up the call stack.