[Truffle] Ensure MatchData's source is a private copy.
We store a copy of the source string in our MatchData instances to make it easier to perform some operations in Ruby. Additionally, we need it to track the taint state. We could store the taint state and the string's rope as separate fields to avoid issues of mutability, but it would preclude writing operations in Ruby without allocating a new string for each such operation. Duping the strings seems like the safest option currently, but this should be revisited.