Define module_function singleton method as proper clone.
We used to use WrapperMethod to carry the alternate implementation
class and visibility for the singleton method. That type of method
does not pass impl class through to the contained method, which
resulted in the eventual "implementer search" from trying to find
an implementer of the module, rather than doing a simple
superclass of the module's singleton. Because the module is
obviously not found in its own singleton class's hierarchy, that
search produced null, and we blindly proceeded to attempt to call
the super method on null.

This now uses a proper clone of the original method for the
singleton version, with its own implClass that propagates into
the implementer search, and super works properly up the singleton
hierarchy.

Fixes #3708.