Don't double-search for requires.
requireCommon and smartLoadInternal had almost the same logic to
check the load against others in progress and do the eventual
search for the file, and the former caller the latter. I added
the one item from rC that was missing from sML (the profile check
for requireability) and removed rC completely.