Make all IR methods cloneable and eliminate trivial dup impls.
This is part of fixes for #3869. Duping methods by reconstructing
them causes them to have a new serial number, which is what we are
using to determine method equality (and not much else). By
changing dup to use clone here, Ruby methods transplanted from
a module to the same module will still be == and eql?.