[Truffle] Simplified the guard checking if a substring would simply be the same as its child.
The old calculation took the offset into account and I think this was a bug. A non-zero offset would trivially ensure the substring and its child were not the same. Since the byte length of the substring is bounded by the byte length of the child (enforced at a higher level), we can reduce the check to simply seeing if the substring has the same byte length as the child.