[Truffle] Reworked String#+ with a RopeBuffer receiver to avoid an unnecessary memory allocation and copy.
The previous implementation duped the RopeBuffer (1 allocation and copy) and then let the MakeConcatRope node expand the RopeBuffer's ByteList by appending the other string's nodes. Unless the other string was empty, this would force the ByteList's backing store to grow (second allocation and copy to temp buffer). This new implementation allocates the resulting ByteList immediately and just copy's the receiver and argument bytes into it.