Fix JI test expecting NativeException-wrapping from constructors.
In 5d057d24 I modified behavior for Java constructors by
re-raising rather than wrapping in NativeException any exceptions
they throw. The NativeException behavior was a relic of a bygone
era, and normal Java methods have not thrown it for several years.
This commit fixes a test that still expected all constructor-
originated exceptions to be wrapped in NativeException; it now
expects the true exception to be raised and asserts accordingly.