[Truffle] Keep track of default_external and default_internal encondings in EncodingManager.
A side effect of this change is we're no longer updating the JRuby runtime's default_external and default_internal values. As a result, we can't use methods from the JRuby runtime that rely on those values being up-to-date. Since RubyString#inspect is one such method, we have to move that back out to Ruby (String#inspect). This code has ping-ponged a few times between a Ruby and Java implementation. It appears whatever issue forced the move back to Java (i.e., RubyString#inspect) has been addressed.