No unsafe gets from closures; parent may not be jitted.
This is a "fix" for #4235 in that it prevents closures from ever
blindly reading closed-over variables that have not been
initialized to nil. This may happen if the block jits independent
of the method, in which case the method will continue to be run
through null-checking interpreter logic, so the closure code
cannot guarantee all values are non-null on first read.

This is only a "fix" in quotes because it completely disables the
fast path optimizaton for all non-method JIT.

Note that I did add specialized "OrNil" paths to the generated
DynamicScope subclasses, so that should at least inline to a field
plus null check.