Fixes #3326. JRuby 9.0.1.0 vs. MRI 2.2.2 define_method inconsistency
Problem was when we switched from arity to signature we added a new
call and we passed the block arg passed to the symbol proc rather
than the symbol proc itself.  Getting this to trigger did not really
fit into any standard ruby behavior so I made a regression spec for
it.  My guess is we only internally call the broken Java method if
the block is set in an uncommon way (although I am not entirely
sure what that is...but define_method will cause it to happen).