Catch LinkageError during DynamicScope generation.
When multiple threads are generating the same DynamicScope class, both
threads may try to generate the same class at the same time. When the second
thread hits defineClass, it throws a LinkageError becuase the class has already
been defined.

With this change we catch this exception and call loadClass to get the class that
the previous thread created.