Disable notifying travis-rubies of passing [skip ci]
As travis-rubies now uses 3 Mac jobs to create archives for various OS
releases.

This is a bit wasteful if multiple builds pass in a short period.

Instead, Travis CI is now running a nightly build of jruby-head
(around 22:30 UTC). This will happen regardless of the state of the
`master` branch.