[Truffle] Removed the `character_ascii_p` primitive.
The only two places this primitive was being used was in String#chop! and String#chomp!. In both cases, the character's codepoint was already being inspected, so checking if it also was in the valid US-ASCII codepoint range was redundant and a bit misleading. There is a bug here in that non-ASCII-compatible encodings may use values other than 0x0A and 0xAD for LF and CR, but the existing code didn't guard against that. Moreover, MRI doesn't guard against that either, so any fix there would come at the expense of breaking compatibility with MRI. As it stands, the only such encodings in MRI are dummy encodings, which aren't fully supported anyway.