Fixes #4342. zsuper does not cope with n repeated _ well.
Basically all recv_* instrs will now check to see if they are assigning
to _ or not.  On first occurrence of _ we will actually assign to it like
an ordinary lvar.  On any additional parameter which uses the name we will
assign it to a temporary variable.  If that scope happens to have a zsuper
then those additional occurrences are not dead.  If not they fall out as
we optimize.