Simplify processing logic of a single incoming argument in a proc to spread that argument if there are multiple required arguments.
In truth, I was hoping to move away from ever using arityValue in backend
logic past arity() itself.  In fact required() (+ 1 if reqdkwarg) I think
is what we really care about at this point but for the lack of a better name
this is '(arityValue < -1 || arityValue > 1)' atm.  Another reason for not
going further is that in lambda processing we have very similar but slightly
different conditional.  I think this check is right as it says: we have at
least 2 parameters which must be filled and we are only passing one...try
and spread that arg.

The second part was to expost toAry() to be public.

I have pretty reasonable confidence level that the convoluted form-fit we have
before boils to this.  It also makes logical sense as well based on how Ruby
spreads arguments.