Remove unsafe-mock from our build altogether.
This commit replaces the unsafe-mock dependency -- which caused
various build problems living in boot classpath -- with a new
library that just shims the three Unsafe methods added in Java 8:
fullFence, loadFence, and storeFence. This allows us to eliminate
all calls to those methods from JRuby proper as well as the build
tricks.

Fixes #4103.