Disable setAccessible in JI when on Java 9.
Java 9's module system requires us to do a bit more work to
set methods accessible within java.lang and other modules. If we
fail to do this, the method raises InaccessibleObjectException,
a Java 9-specific exception we can't catch in Java 8 code. For the
moment we will disable setAccessible use on Java 9 so JRuby can
start up and run correctly.