Lazily create now-unused idTest handle and share method reference.
The handle here was originally intended for use in checking that
the target module for a constant cache is the same as when the
value was cached. However the code that used it is no longer
active, but we still created these handles for every module,
class, singleton class, included module, etc in the system.

This commit makes the acquisition of this handle lazy, and also
modifies it to reuse the direct handle reference to the test
method rather than recreating that every time. This should reduce
the number of method handles in flight.

For #4391.