Fixes #4154.  Wrong line number reported in backtrace in hash creation.
I took a somewhat minimal approach to fixing this in that all calls can
potentially raise in execution so I potentially add linenum instrs when
building any type of call node.  In theory anything with a side-effect
could cause this like a special global etc...  This will hit high 90%
of the issue at least.  I am also wondering if this will regress anything
so it is a good tester before I make the AST know about whether a subtree
contains a potential side-effecting node (which is the longer term fix).