Re-push the frame, class, file, and line from the thread block.
When a non-local flow event bubbles out of a thread, it lazily
triggers a LocalJumpError to be created, along which path a hook
event for "raise" fires. Unfortunately this tries to access the
current frame and stack trace information, which are already
popped off the stack by the thread's proc.

This change re-pushes that frame and stack trace info around the
creation of the exception, so that the hook has appropriate info
available.

Fixes #3781