Propagate original bytes (ByteList) from parser to symbols.
The first fix here, which fixes #4564, stopped the Symbol operand
from modifying an existing symbol with the same bytes to have a
new Encoding. This was broken because sometimes the encoding it
chose, along some paths, was the wrong one, and it also modified
a shared ByteList causing sym.to_s strings elsewhere to suddenly
change encoding.

Fixing that led to another set of issues that required fixing.

Our Symbol logic in the AST and IR was trying to shove the
original parsed bytes through as a java.lang.String plus encoding.
Unfortunately the parser was providing that string as ISO-8859-1
bytes and then attempting to encode the string using the provided
encoding, resulting in bogus multibyte characters getting into IR.
The fix for #4564 removed broken logic in IR that forced an
existing symbol to have a new encoding, so the broken bytes got
into the table and stayed there.

With this change, Symbols never change encoding, *most* paths to
create them provide the original ByteList, and the remaining
java.lang.String paths encode to bytes properly.

Fixes #4564 (and probably other cases).