Fix most ObjectSpace-triggered Marshal spec failures.
The problem here was that when using ObjectSpace (or object_id,
or FFI) we need to insert some hidden state into objects. This
state goes into the variable table, but does not get marshaled.
However, if we had both reserved space for instance variables and
for one of these hidden variables, we always would proceed to
marshal the object as though it might have instance variables. If
none of the non-hidden variables had been set, we marshaled the
object as having instance variables, but of zero length. This
broke several marshal specs, since objects that would otherwise
have no instance variable were marshaled as such.

This fix uses a second check against the variable list to ensure
that there are real variables to marshal (or an encoding variable)
before proceeding to emit the instance variable sigil and list.