Accurately track the three hidden vars in the var manager.
This is not directly used other than to provide a more accurate
guess as to whether there are any non-hidden variable in use by
a given type. 0b6e3c00 fixes the original issue (marshal dumping
zero-length instance var lists).

The accessors of the lazy fields are now deprecated. Use the read
or write calls to get an accessor for the hidden fields.