Fixes #4662. LoadError with compiled multi-byte constant.
The original code was using String.length() then saving getBytes() but length()
reports characters and not bytes so multi-byte strings were not getting saved
with the correct length.

More commentary in the issue since even with this fix we are still broken in a
known but more fundamental way (and we have since the beginning of time).