Reinstate a non-indy const lookup path for Android.
This is not a straight reversion of the caching commit below,
since there were modifications to how constants bound after that
change.

Note that we still need to restore constant caching for non-indy
paths, or otherwise make it easier to say that Java 7+ JRuby will
use indy for some things without compile.invokedynamic, but
disable those features on platforms that don't support indy.

Revert "Always use invokeynamic to lookup and cache constants."

This reverts commit e03c4f16581a61259d4d041760215ac329bae278.