Start of using symbols as identifiers.  Only local variable names and only for static scope.  IRBuilder has not been changed to use them.
The main desire for Symbols is we maintain a single pool of them.  When we
were using String we would intern to get similar benefit.  Current parser is
making lots and lots of ByteLists.  This will end up having a memory impact
(one I have not actually measured) but by using Symbols we will be back to
a String.intern amount.

Symbols also conceptually is what all reflective APIs already want.