Simplify case/when eqq logic and enlist in normal invocation.
Most case/when will have neither an undefined case value (for bare
case statements) nor multiple values for a given when (splatted)
so in those cases we can simply do a plain eqq call + isTrue. The
unusual cases will still go to the old isEQQ logic, so there may
be value in optimizing that path later.

The main logic in isEQQ is the processing of a splatted array of
when values in sequence, and a small bit of logic that simply
checks truthiness of the when value for a bare case statement.

This change reduces the bytecode for all case/when, and by more
than half for the typical form. It also allows the === call to be
optimized with invokedynamic like a normal call.