Remove BEQ.
BEQ appeared to only get used by case/when statements where a
when has a literal array as a first element. This meant that for
when with a literal array, we were calling #== instead of #===.

I could not reproduce this behavior in MRI, and the following
code shows that we are doing something wrong.

```ruby
class Array
  def ===(other)
    p :===
    self == other
  end
end

case x
when []
end
```

MRI prints out :=== while JRuby does not.

This logic apears to be from when we used to process whens with
multiple arguments differently. Today, those cases get built as
separate conditions, and I believe this removed code is no longer
valid.