Reinstate the cat path using the full string.
Previously this logic used cat but without the code range
requiring it to be re-scanned later. Since the literal string
components of a dstr now are always the same frozen string
instance, we can preserve the code range and call the version
of cat that makes use of it. This also skips the type checks in
the untyped append path (used for the dynamic portions of a dstr).