Synchronize bytecode JIT to avoid concurrency issues.
This effectively means the background JIT will usually be limited
to one thread operating at a time. How much this will slow down
or bottleneck the JIT is not clear but this fix will let us know
if we need a deeper look at synchronization of the JIT and IR
structures modified at JIT time.

Part of work for #4739.