Include hierarchy of prepends when adding method names from mod.
Fixes #4723

Method searching for calls used overridden logic in the included
module wrapper, but we did not do the same logic when
traversing all methods to add them to a reflective method list
(such as instance_methods). This patch adds overridden logic in
the wrapper to walk the prepends just as in the call site case.