Fix block escaping logic in interpreter one arg block calls.
The "full" interpreter inlines some call logic to avoid extra
dispatch. In this case, the logic it inlined did not include
logic to only escape the block if it were passed as a literal
block. Instead, it escaped it all the time. Interestingly, the
actual instruction (used by the startup interpreter) was wrong
in the opposite way: it never escaped the block.

Both places have been fixed to do the right thing and comments
added to help ensure they remain the same.

Fixes #4841.