Fixes #4865. OOM due to unbounded rescuePCs growth.
The comment in this commit hopefully spells out how the solution works but one
tl;dr is we cannot guarantee start and end exception region instructions will
occur in the presence of branches or jumps.  The workaround is to clean up the
stack whenever we decide to add a new exception region.

This had a small side-benefit of removing a hack we must have added when seeing
a jump-like instr just doing a pop and hoping for the best.