Don't unconditionally emit debug level info from RACC
In
https://github.com/jruby/jruby/commit/3fe1cc2353e8b051a7541d3cb63d5210e99d8539,
jruby's version of racc/parser.rb was brought in line with MRI
racc/parser.rb. This brought along a change to the yyparse method which
changed the last argument from false to true. This causes the java racc
parser implementation to unconditionally emit very noisy debug output
whenever yyparse is called, where previously it would not emit such
output. Passing true is safe for the pure ruby parser, as it ignores the
argument entirely and does nothing with it. It is also safe for the
native c parser, as that parser only emits debug info if it is compiled
with DEBUG set. Tenderlove's version of racc disabled debug by default
in
https://github.com/tenderlove/racc/commit/c5359ab04b5ad858d59664fbe288e9cc7fb1c27f,
but it looks like MRI never picked up that change.
For a time, true was passed and debugging info was not emitted (between
9.1.9.0 and 9.1.12.0). This was because in the merge to bring
racc/parser.rb in line with MRI, the loading of the java cparse library
no longer happened, so the pure ruby version of the library was used
which ignores the debug parameter entirely. This changed in
https://github.com/jruby/jruby/commit/baeaad12f32842a5b714c463558feeb2f7b34c0b
when the cparse-java.jar was being loaded again for jruby.