Re-add an mbean for monitoring cache invalidations.
This is for #4384.

For this initial version, methods and constants only have a single
value indicating validation count globally. For constants, this
will aggregate all constant names (which are invalidated
separately). For methods, this will aggregate all class hierarchy
invalidations, which will frequently mean a single invalidation
event will increment this value many times, depending on the size
of the hierarchy below it. This latter point could be improved
but we don't really have a clear place to say "this is the top
invalidation" at the moment.